fix dependecy cycle
