+++ src/assets/css/style.css
@@ -87,7 +87,7 @@ video {
   margin: 0;
   padding: 0;
   border: 0;
-  font-size: 8px;
+  font-size: 16px;
   vertical-align: baseline;
 }
