--- src/renderSideBar.js
@@ -1,72 +1,0 @@
-import render from './render';
-import Project from './project';
-
-// RENDER SIDEBAR
-const renderSideBar = (projects) => {
-  // CLEAR SIDEBAR TO RENDER UPDATED LIST
-  const sideBar = document.createElement('div');
-  sideBar.classList.add('side-bar');
-  sideBar.innerHTML = '';
-
-  // RENDER EACH PROJECT NAME IN SIDEBAR
-  projects.forEach((project, index) => {
-    const projectBtn = document.createElement('button');
-    projectBtn.classList.add('project-btn');
-    projectBtn.textContent = project.name;
-    projectBtn.addEventListener('click', () => {
-      render(project);
-    });
-
-    if (index !== 0) {
-      const projectDeleteBtn = document.createElement('button');
-      projectDeleteBtn.classList.add('project-delete-btn');
-      projectDeleteBtn.textContent = ' X ';
-      projectDeleteBtn.addEventListener('click', (event) => {
-        event.stopPropagation();
-        projects.splice(index, 1);
-        render();
-      // storage.updateLocalStorage();
-      });
-
-      projectBtn.appendChild(projectDeleteBtn);
-    }
-
-    sideBar.appendChild(projectBtn);
-  });
-
-  //  NEW PROJECT FORM
-  const newProjectForm = document.createElement('div');
-  newProjectForm.classList.add('new-project-wrapper', 'hidden');
-
-  const inputField = document.createElement('input');
-  inputField.classList.add('new-project-name');
-
-  const saveBtn = document.createElement('button');
-  saveBtn.classList.add('save-project-btn');
-  saveBtn.textContent = 'save';
-  saveBtn.addEventListener('click', () => {
-    projects.push(new Project(inputField.value));
-    render(projects[projects.length - 1]);
-    // storage.updateLocalStorage();
-  });
-
-  newProjectForm.appendChild(inputField);
-  newProjectForm.appendChild(saveBtn);
-
-  sideBar.appendChild(newProjectForm);
-
-  // ADD PROJECT BUTTON
-  const addProjectBtn = document.createElement('button');
-  addProjectBtn.classList.add('add-project-btn');
-  addProjectBtn.textContent = 'Add Project';
-
-  addProjectBtn.addEventListener('click', () => {
-    newProjectForm.classList.toggle('hidden');
-  });
-
-  sideBar.appendChild(addProjectBtn);
-
-  return sideBar;
-};
-
-export default renderSideBar;
