--- src/todoForm.js
@@ -1,107 +1,0 @@
-import render from './render';
-import Todo from './todo';
-
-const getDateString = (date) => {
-  date.setMinutes(date.getMinutes() - date.getTimezoneOffset());
-  return date.toJSON().slice(0, 10);
-};
-
-class TodoForm {
-  constructor(...params) {
-    [this.todoObj, this.project] = params;
-    // let title, description, dueDate, priority;
-    const {
-      title = '', description = '', dueDate = getDateString(new Date()), priority = '',
-    } = this.todoObj;
-
-    this.titleField = document.createElement('input');
-    this.descriptionField = document.createElement('textarea');
-    this.dueDateField = document.createElement('input');
-    this.priorityField = document.createElement('select');
-
-    this.titleField.value = title;
-    this.descriptionField.value = description;
-    this.dueDateField.value = dueDate;
-    this.priorityField.value = priority;
-
-    this.priority = priority;
-    // this.project = project;
-  }
-
-  render() { // define type of button action
-    this.todoFormContainer = document.createElement('div');
-
-    // STYLING FORM ELEMENTS
-    this.todoFormContainer.classList.add('new-todo-wrapper');
-    this.titleField.classList.add('new-todo-title');
-    this.titleField.setAttribute('placeholder', 'Title');
-    this.descriptionField.classList.add('new-todo-description');
-    this.descriptionField.setAttribute('placeholder', 'Description');
-    this.dueDateField.classList.add('new-todo-due-date');
-    this.dueDateField.setAttribute('type', 'date');
-    this.priorityField.classList.add('new-todo-priority');
-    this.priorityField.setAttribute('name', 'new-todo-priority');
-
-    let [lowSelected, midSelected, highSelected] = ['', '', ''];
-
-    switch (this.priority) {
-      case 'high':
-        highSelected = 'selected = "selected"';
-        break;
-      case 'mid':
-        midSelected = 'selected = "selected"';
-        break;
-      default:
-        lowSelected = 'selected = "selected"';
-        break;
-    }
-
-    this.priorityField.innerHTML = `
-    <option value="low" ${lowSelected}>low</option>
-    <option value="mid" ${midSelected}>mid</option>
-    <option value="high" ${highSelected}>high</option>
-    `;
-
-    // SAVE BTN
-    this.saveBtn = document.createElement('button');
-    this.saveBtn.classList.add('save-todo-btn');
-    this.saveBtn.textContent = 'save';
-    this.saveBtn.addEventListener('click', () => {
-      const newTodoObj = {
-        title: this.titleField.value,
-        description: this.descriptionField.value,
-        dueDate: this.dueDateField.value,
-        priority: this.priorityField.value,
-      };
-
-      // CHECK IF FIELD IS EMPTY
-      const inputIsValid = Object.entries(newTodoObj).every((field) => {
-        if (field[1] === '') {
-          alert(`The field ${field[0]} can't be empty!`);
-          return false;
-        }
-
-        return true;
-      });
-
-      if (inputIsValid) {
-        if (this.todoObj instanceof Todo) {
-          this.todoObj.updateTodo(newTodoObj);
-        } else {
-          this.project.addTodo(newTodoObj);
-        }
-
-        render(this.project);
-      }
-    });
-
-    this.todoFormContainer.appendChild(this.titleField);
-    this.todoFormContainer.appendChild(this.descriptionField);
-    this.todoFormContainer.appendChild(this.dueDateField);
-    this.todoFormContainer.appendChild(this.priorityField);
-    this.todoFormContainer.appendChild(this.saveBtn); // BUTTON!
-
-    return this.todoFormContainer;
-  }
-}
-export default TodoForm;
